make changes after discussion about PR
